extract AdvancedSearch subcomponent: DesktopSearchWithSideTabs
